+++ Micon.CMS/Controllers/PageTemplateController.cs
@@ -1,4 +1,5 @@
 ﻿using Micon.CMS.Models;
+using Micon.CMS.Models.Api;
 using Micon.CMS.Repositories;
 using Micon.CMS.Services;
 using Micon.CMS.Library.Models.Form;
@@ -21,6 +22,8 @@ namespace Micon.CMS.Controllers
         private readonly IViewComponentHelper _viewComponentHelper;
         private readonly ComponentSlotAnalyzerService _slotAnalyzerService;
         private readonly IComponentCacheService _componentCacheService;
+        private readonly ComponentTreeService _treeService;
+        private readonly WorkspaceService _workspaceService;
 
         public PageTemplateController(
             IPageTemplateRepository pageTemplateRepository,
@@ -29,7 +32,9 @@ namespace Micon.CMS.Controllers
             IPageTemplateWorkspaceRepository workspaceRepository,
             IViewComponentHelper viewComponentHelper,
             ComponentSlotAnalyzerService slotAnalyzerService,
-            IComponentCacheService componentCacheService)
+            IComponentCacheService componentCacheService,
+            ComponentTreeService treeService,
+            WorkspaceService workspaceService)
         {
             _pageTemplateRepository = pageTemplateRepository;
             _componentRepository = componentRepository;
@@ -38,6 +43,8 @@ namespace Micon.CMS.Controllers
             _viewComponentHelper = viewComponentHelper;
             _slotAnalyzerService = slotAnalyzerService;
             _componentCacheService = componentCacheService;
+            _treeService = treeService;
+            _workspaceService = workspaceService;
         }
         public async Task<IActionResult> Index(CancellationToken cancellation)
         {
@@ -62,144 +69,29 @@ namespace Micon.CMS.Controllers
 
         public async Task<IActionResult> Edit(Guid Id, CancellationToken cancellationToken)
         {
-            Console.WriteLine($"Edit action called with Id: {Id}");
-
             var pageTemplate = await _pageTemplateRepository.GetByIdAsync(Id, cancellationToken);
             if (pageTemplate == null)
             {
-                Console.WriteLine("PageTemplate not found");
                 return RedirectToAction(nameof(Index));
             }
 
-            Console.WriteLine($"PageTemplate found: {pageTemplate.Name}");
+            // Workspace を取得または作成
+            var workspace = await _workspaceService.GetOrCreateWorkspaceAsync(Id, pageTemplate, cancellationToken);
+            var initialTree = _workspaceService.GetTreeFromWorkspace(workspace);
 
-            // ワークスペースを取得、存在しなければ作成
-            var workspace = await _workspaceRepository.GetByPageTemplateIdAsync(Id, cancellationToken);
-
-            if (workspace == null)
-            {
-                Console.WriteLine("Workspace not found, creating new one");
-                // 新しいワークスペースを作成
-                var initialTree = new List<ComponentTreeNode>();
-
-                // 元の PageTemplate に ComponentRelation がある場合、ツリーとして保存
-                if (pageTemplate.ComponentRelationId.HasValue)
-                {
-                    var hierarchy = await _pageTemplateRepository.GetComponentHierarchy(pageTemplate, cancellationToken);
-                    initialTree = await BuildComponentTreeFromHierarchy(hierarchy, cancellationToken);
-                }
-
-                workspace = new PageTemplateWorkspace
-                {
-                    PageTemplateId = Id,
-                    WorkspaceData = System.Text.Json.JsonSerializer.Serialize(initialTree),
-                    CreatedByUserId = null,
-                    UpdatedAt = DateTime.UtcNow
-                };
-
-                workspace = await _workspaceRepository.CreateAsync(workspace, cancellationToken);
-                Console.WriteLine($"Workspace created with {initialTree.Count} root nodes");
-            }
-            else
-            {
-                Console.WriteLine($"Workspace found: {workspace.Id}, Data length: {workspace.WorkspaceData.Length}");
-            }
-
-            // コンポーネント情報とスロット情報を取得（キャッシュから）
+            // コンポーネント情報を取得
             var components = await _componentCacheService.GetCachedComponentsAsync();
             var componentsWithSlots = await BuildComponentsWithSlots(components, cancellationToken);
             ViewBag.Components = componentsWithSlots;
-            ViewBag.WorkspaceId = workspace.Id;
+            ViewBag.ComponentHierarchy = initialTree;
 
-            // ワークスペースのツリーをデシリアライズして表示
-            var workspaceTree = System.Text.Json.JsonSerializer.Deserialize<List<ComponentTreeNode>>(workspace.WorkspaceData) ?? new List<ComponentTreeNode>();
-            Console.WriteLine($"Deserialized workspace tree with {workspaceTree.Count} root nodes");
-            Console.WriteLine($"Workspace data being sent to view: {System.Text.Json.JsonSerializer.Serialize(workspaceTree)}");
-
-            ViewBag.ComponentHierarchy = workspaceTree;
-
-            // ワークスペースのツリーからPageComponentViewModelを構築してプレビュー用に渡す
-            var previewViewModel = await BuildPageComponentViewModelFromWorkspace(workspaceTree, pageTemplate.Id, cancellationToken);
+            // プレビュー用ViewModel を構築
+            var previewViewModel = await BuildPageComponentViewModelFromWorkspace(initialTree, pageTemplate.Id, cancellationToken);
             ViewBag.PreviewViewModel = previewViewModel;
 
             return View(pageTemplate);
         }
 
-        private async Task<List<ComponentTreeNode>> BuildComponentTreeFromHierarchy(List<Models.ComponentHierarchy> hierarchy, CancellationToken cancellationToken)
-        {
-            if (hierarchy == null || !hierarchy.Any())
-            {
-                return new List<ComponentTreeNode>();
-            }
-
-            var rootNodes = hierarchy.Where(h => h.ParentId == null || h.ParentId == Guid.Empty).ToList();
-            var result = new List<ComponentTreeNode>();
-
-            foreach (var root in rootNodes)
-            {
-                result.Add(await BuildNodeTree(root, hierarchy, cancellationToken));
-            }
-
-            return result;
-        }
-
-        private async Task<ComponentTreeNode> BuildNodeTree(Models.ComponentHierarchy hierarchyItem, List<Models.ComponentHierarchy> allHierarchy, CancellationToken cancellationToken)
-        {
-            var node = new ComponentTreeNode
-            {
-                ComponentId = hierarchyItem.ChildComponentId.ToString(),
-                ComponentName = hierarchyItem.ChildComponentName,
-                PackageId = hierarchyItem.ChildComponentPackageId.ToString(),
-                SlotName = hierarchyItem.SlotName,
-                Children = new List<ComponentTreeNode>()
-            };
-
-            var childItems = allHierarchy.Where(h => h.ParentComponentId == hierarchyItem.ChildComponentId).ToList();
-            foreach (var child in childItems)
-            {
-                node.Children.Add(await BuildNodeTree(child, allHierarchy, cancellationToken));
-            }
-
-            return node;
-        }
-
-        private Micon.CMS.Library.Models.Form.PageComponentViewModel BuildComponentViewModelTree(List<Models.ComponentHierarchy> hierarchy)
-        {
-            if (hierarchy == null || !hierarchy.Any())
-            {
-                return new Micon.CMS.Library.Models.Form.PageComponentViewModel();
-            }
-
-            var lookup = hierarchy.ToDictionary(
-                h => h.ChildId,
-                h => new Micon.CMS.Library.Models.Form.PageComponentViewModel
-                {
-                    ComponentId = h.ChildComponentId,
-                    ComponentName = h.ChildComponentName,
-                    SlotName = h.SlotName,
-                    PackageId = h.ChildComponentPackageId
-                });
-
-            var rootNode = new Micon.CMS.Library.Models.Form.PageComponentViewModel
-            {
-                ComponentName = "Root",
-                SlotName = "Main"
-            };
-
-            foreach (var item in hierarchy)
-            {
-                if (item.ParentId.HasValue && lookup.ContainsKey(item.ParentId.Value))
-                {
-                    lookup[item.ParentId.Value].Children.Add(lookup[item.ChildId]);
-                }
-                else
-                {
-                    rootNode.Children.Add(lookup[item.ChildId]);
-                }
-            }
-
-            return rootNode.Children.Count == 1 ? rootNode.Children.First() : rootNode;
-        }
 
         [HttpPost]
         public async Task<IActionResult> InitializeComponentRelation(Guid Id, CancellationToken cancellationToken)
@@ -221,56 +113,27 @@ namespace Micon.CMS.Controllers
         {
             try
             {
-                Console.WriteLine($"AddComponent called with: {System.Text.Json.JsonSerializer.Serialize(request)}");
-
-                if (request == null)
+                if (request == null || !Guid.TryParse(request.PageTemplateId, out var pageTemplateId) || !Guid.TryParse(request.ComponentId, out var componentId))
                 {
-                    return BadRequest("Request body is required");
+                    return BadRequest("Invalid request parameters");
                 }
 
-                if (!Guid.TryParse(request.PageTemplateId, out var pageTemplateId))
-                {
-                    Console.WriteLine($"Invalid pageTemplateId: {request.PageTemplateId}");
-                    return BadRequest("Invalid pageTemplateId format");
-                }
-
-                if (!Guid.TryParse(request.ComponentId, out var componentId))
-                {
-                    Console.WriteLine($"Invalid componentId: {request.ComponentId}");
-                    return BadRequest("Invalid componentId format");
-                }
-
-                Console.WriteLine($"Looking for PageTemplate: {pageTemplateId}");
                 var pageTemplate = await _pageTemplateRepository.GetByIdAsync(pageTemplateId, cancellationToken);
                 if (pageTemplate == null)
                 {
-                    Console.WriteLine("PageTemplate not found");
                     return NotFound("PageTemplate not found");
                 }
 
-                Console.WriteLine($"Looking for Component: {componentId}");
                 var cachedComponents = await _componentCacheService.GetCachedComponentsAsync();
                 var component = cachedComponents.FirstOrDefault(c => c.Id == componentId);
                 if (component == null)
                 {
-                    Console.WriteLine("Component not found in cache");
                     return BadRequest("Component not found");
                 }
 
-                // Workspace を取得
-                Console.WriteLine($"Looking for Workspace for PageTemplate: {pageTemplateId}");
-                var workspace = await _workspaceRepository.GetByPageTemplateIdAsync(pageTemplateId, cancellationToken);
-                if (workspace == null)
-                {
-                    Console.WriteLine("Workspace not found");
-                    return NotFound("Workspace not found");
-                }
-
-                Console.WriteLine($"Workspace found: {workspace.Id}, Data length: {workspace.WorkspaceData.Length}");
-
-                // ワークスペースのツリーをデシリアライズ
-                var tree = System.Text.Json.JsonSerializer.Deserialize<List<ComponentTreeNode>>(workspace.WorkspaceData) ?? new List<ComponentTreeNode>();
-                Console.WriteLine($"Deserialized tree with {tree.Count} root nodes");
+                // Workspace を取得または作成
+                var workspace = await _workspaceService.GetOrCreateWorkspaceAsync(pageTemplateId, pageTemplate, cancellationToken);
+                var tree = _workspaceService.GetTreeFromWorkspace(workspace);
 
                 // 新しいノードを作成
                 var newNode = new ComponentTreeNode
@@ -285,55 +148,32 @@ namespace Micon.CMS.Controllers
                 // ツリーに追加
                 if (request.ParentPath == null || request.ParentPath.Count == 0)
                 {
-                    Console.WriteLine("Adding node as root");
                     tree.Add(newNode);
                 }
                 else
                 {
-                    Console.WriteLine($"Adding node as child at path: {string.Join(",", request.ParentPath)}");
                     var parent = GetNodeByPath(tree, request.ParentPath);
-                    if (parent != null)
+                    if (parent == null)
                     {
-                        var targetSlot = newNode.SlotName ?? "Main";
-                        if (parent.Children.Any(c => string.Equals(c.SlotName ?? "Main", targetSlot, StringComparison.OrdinalIgnoreCase)))
-                        {
-                            Console.WriteLine("Parent slot already occupied");
-                            return BadRequest("指定したスロットには既にコンポーネントが配置されています。先に削除してください。");
-                        }
-
-                        parent.Children.Add(newNode);
-                        Console.WriteLine("Child node added successfully");
-                    }
-                    else
-                    {
-                        Console.WriteLine("Parent node not found");
                         return BadRequest("Parent node not found");
                     }
+
+                    var targetSlot = newNode.SlotName ?? "Main";
+                    if (parent.Children.Any(c => string.Equals(c.SlotName ?? "Main", targetSlot, StringComparison.OrdinalIgnoreCase)))
+                    {
+                        return BadRequest("指定したスロットには既にコンポーネントが配置されています。先に削除してください。");
+                    }
+
+                    parent.Children.Add(newNode);
                 }
 
-                // ワークスペースを更新して保存
-                workspace.WorkspaceData = System.Text.Json.JsonSerializer.Serialize(tree);
-                workspace.UpdatedAt = DateTime.UtcNow;
-                Console.WriteLine($"Saving workspace with {tree.Count} root nodes");
-                await _workspaceRepository.UpdateAsync(workspace, cancellationToken);
-                Console.WriteLine("Workspace saved successfully");
+                // Workspace を保存
+                await _workspaceService.SaveTreeToWorkspaceAsync(workspace, tree, cancellationToken);
 
-                // 更新後のツリーと追加コンポーネント情報を返却
-                return Ok(new
-                {
-                    tree,
-                    addedComponent = new
-                    {
-                        componentId = component.Id.ToString(),
-                        componentName = component.Name,
-                        packageId = component.PackageId.ToString(),
-                        slotName = request.SlotName ?? "Main"
-                    }
-                });
+                return Ok(new { tree });
             }
             catch (Exception ex)
             {
-                Console.WriteLine($"Exception in AddComponent: {ex}");
                 return StatusCode(500, $"Error: {ex.Message}");
             }
         }
@@ -343,37 +183,28 @@ namespace Micon.CMS.Controllers
         {
             try
             {
-                if (request == null)
+                if (request == null || !Guid.TryParse(request.PageTemplateId, out var pageTemplateId) || request.Path == null || request.Path.Count == 0)
                 {
-                    return BadRequest("Request body is required");
+                    return BadRequest("Invalid request parameters");
                 }
 
-                if (!Guid.TryParse(request.PageTemplateId, out var pageTemplateId))
+                var pageTemplate = await _pageTemplateRepository.GetByIdAsync(pageTemplateId, cancellationToken);
+                if (pageTemplate == null)
                 {
-                    return BadRequest("Invalid pageTemplateId format");
+                    return NotFound("PageTemplate not found");
                 }
 
-                if (request.Path == null || request.Path.Count == 0)
-                {
-                    return BadRequest("Path is required");
-                }
-
-                var workspace = await _workspaceRepository.GetByPageTemplateIdAsync(pageTemplateId, cancellationToken);
-                if (workspace == null)
-                {
-                    return NotFound("Workspace not found");
-                }
-
-                var tree = System.Text.Json.JsonSerializer.Deserialize<List<ComponentTreeNode>>(workspace.WorkspaceData) ?? new List<ComponentTreeNode>();
+                // Workspace を取得または作成
+                var workspace = await _workspaceService.GetOrCreateWorkspaceAsync(pageTemplateId, pageTemplate, cancellationToken);
+                var tree = _workspaceService.GetTreeFromWorkspace(workspace);
 
                 if (!RemoveNodeFromTree(tree, request.Path))
                 {
                     return BadRequest("Target node not found");
                 }
 
-                workspace.WorkspaceData = System.Text.Json.JsonSerializer.Serialize(tree);
-                workspace.UpdatedAt = DateTime.UtcNow;
-                await _workspaceRepository.UpdateAsync(workspace, cancellationToken);
+                // Workspace を保存
+                await _workspaceService.SaveTreeToWorkspaceAsync(workspace, tree, cancellationToken);
 
                 return Ok(new { tree });
             }
@@ -529,14 +360,9 @@ namespace Micon.CMS.Controllers
         {
             try
             {
-                if (request == null)
+                if (request == null || !Guid.TryParse(request.PageTemplateId, out var pageTemplateId))
                 {
-                    return BadRequest("Request body is required");
-                }
-
-                if (!Guid.TryParse(request.PageTemplateId, out var pageTemplateId))
-                {
-                    return BadRequest("Invalid pageTemplateId format");
+                    return BadRequest("Invalid request");
                 }
 
                 var pageTemplate = await _pageTemplateRepository.GetByIdAsync(pageTemplateId, cancellationToken);
@@ -545,36 +371,28 @@ namespace Micon.CMS.Controllers
                     return NotFound("PageTemplate not found");
                 }
 
-                // Workspace を取得
                 var workspace = await _workspaceRepository.GetByPageTemplateIdAsync(pageTemplateId, cancellationToken);
-                if (workspace == null)
-                {
-                    return NotFound("Workspace not found");
-                }
 
-                // ツリーが空の場合は ComponentRelationId をクリア
+                // ツリーが空の場合
                 if (request.Tree == null || request.Tree.Count == 0)
                 {
-                    // 既存の ComponentRelation を削除
                     if (pageTemplate.ComponentRelationId.HasValue)
                     {
                         var existingRelation = await _componentRelationRepository.GetByIdAsync(pageTemplate.ComponentRelationId.Value, cancellationToken);
                         if (existingRelation != null)
                         {
-                            await _componentRelationRepository.DeleteAsync(existingRelation, cancellationToken);
+                            await _treeService.DeleteComponentRelationTree(existingRelation, cancellationToken);
                         }
                     }
 
                     pageTemplate.ComponentRelationId = null;
                     await _pageTemplateRepository.UpdateAsync(pageTemplate, cancellationToken);
-
-                    // ワークスペースを削除
-                    await _workspaceRepository.DeleteAsync(workspace, cancellationToken);
+                    await _workspaceService.DeleteWorkspaceAsync(workspace, cancellationToken);
 
                     return Ok(new { message = "PageTemplate saved successfully (empty tree)" });
                 }
 
-                // 既存の ComponentRelation ツリーを取得
+                // 既存の ComponentRelation を取得
                 ComponentRelation existingRootRelation = null;
                 if (pageTemplate.ComponentRelationId.HasValue)
                 {
@@ -583,7 +401,7 @@ namespace Micon.CMS.Controllers
 
                 // ツリーを差分管理で保存
                 var rootNode = request.Tree[0];
-                var rootRelation = await MergeComponentRelationTree(rootNode, existingRootRelation, null, 0, cancellationToken);
+                var rootRelation = await _treeService.MergeComponentRelationTree(rootNode, existingRootRelation, null, 0, cancellationToken);
 
                 if (rootRelation != null)
                 {
@@ -591,10 +409,10 @@ namespace Micon.CMS.Controllers
                     await _pageTemplateRepository.UpdateAsync(pageTemplate, cancellationToken);
                 }
 
-                // ワークスペースを削除
-                await _workspaceRepository.DeleteAsync(workspace, cancellationToken);
+                // Workspace を削除
+                await _workspaceService.DeleteWorkspaceAsync(workspace, cancellationToken);
 
-                return Ok(new { message = "PageTemplate saved successfully", componentRelationId = rootRelation?.Id });
+                return Ok(new { message = "PageTemplate saved successfully" });
             }
             catch (Exception ex)
             {
@@ -614,358 +432,6 @@ namespace Micon.CMS.Controllers
             }));
         }
 
-        [HttpPost]
-        public async Task<IActionResult> SaveComponentTree([FromBody] SaveComponentTreeRequest request, CancellationToken cancellationToken)
-        {
-            try
-            {
-                if (request == null)
-                {
-                    return BadRequest("Request body is required");
-                }
-
-                if (!Guid.TryParse(request.PageTemplateId, out var pageTemplateId))
-                {
-                    return BadRequest("Invalid pageTemplateId format");
-                }
-
-                var pageTemplate = await _pageTemplateRepository.GetByIdAsync(pageTemplateId, cancellationToken);
-                if (pageTemplate == null)
-                {
-                    return NotFound("PageTemplate not found");
-                }
-
-                // 既存のComponentRelationを削除
-                if (pageTemplate.ComponentRelationId.HasValue)
-                {
-                    var existingRelation = await _componentRelationRepository.GetByIdAsync(pageTemplate.ComponentRelationId.Value, cancellationToken);
-                    if (existingRelation != null)
-                    {
-                        await _componentRelationRepository.DeleteAsync(existingRelation, cancellationToken);
-                    }
-                    pageTemplate.ComponentRelationId = null;
-                }
-
-                // ツリーが空の場合は保存して返す
-                if (request.Tree == null || request.Tree.Count == 0)
-                {
-                    await _pageTemplateRepository.UpdateAsync(pageTemplate, cancellationToken);
-                    return Ok(new { message = "Component tree saved successfully (empty)" });
-                }
-
-                // ツリーを保存（ルートノードのみ）
-                var rootNode = request.Tree[0];
-                var rootRelation = await CreateComponentRelationTree(rootNode, null, 0, cancellationToken);
-
-                if (rootRelation != null)
-                {
-                    pageTemplate.ComponentRelationId = rootRelation.Id;
-                    await _pageTemplateRepository.UpdateAsync(pageTemplate, cancellationToken);
-                }
-
-                return Ok(new { message = "Component tree saved successfully", componentRelationId = rootRelation?.Id });
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, $"Error: {ex.Message}");
-            }
-        }
-
-        private async Task<ComponentRelation> CreateComponentRelationTree(
-            ComponentTreeNode node,
-            Guid? parentId,
-            int order,
-            CancellationToken cancellationToken)
-        {
-            if (!Guid.TryParse(node.ComponentId, out var componentId))
-            {
-                return null;
-            }
-
-            var component = await _componentRepository.GetByIdAsync(componentId, cancellationToken);
-            if (component == null)
-            {
-                return null;
-            }
-
-            var relation = new ComponentRelation
-            {
-                Id = Guid.NewGuid(),
-                ParentId = parentId,
-                ChildId = componentId,
-                SlotName = node.SlotName ?? "Main",
-                Order = order,
-                IsPriority = false
-            };
-
-            await _componentRelationRepository.CreateAsync(relation, cancellationToken);
-
-            // 子ノードを処理
-            if (node.Children != null && node.Children.Count > 0)
-            {
-                int childOrder = 0;
-                foreach (var child in node.Children)
-                {
-                    await CreateComponentRelationTree(child, relation.ChildId, childOrder++, cancellationToken);
-                }
-            }
-
-            return relation;
-        }
-
-        /// <summary>
-        /// 既存のComponentRelationツリーと新しいツリーを比較・マージして差分管理
-        /// PackageId と Name が同じコンポーネントの関係は既存のレコードを再利用
-        /// </summary>
-        private async Task<ComponentRelation> MergeComponentRelationTree(
-            ComponentTreeNode newNode,
-            ComponentRelation existingRelation,
-            Guid? parentId,
-            int order,
-            CancellationToken cancellationToken)
-        {
-            var component = await ExtractComponentAsync(newNode.ComponentId, cancellationToken);
-            if (component == null)
-            {
-                return null;
-            }
-
-            // 既存のRelationが同じ構造か判定
-            var slotName = newNode.SlotName ?? "Main";
-            var isSameStructure = IsRelationUnchanged(existingRelation, component, slotName, parentId, order);
-
-            // 既存の関連を再利用するか、新しいものを作成
-            var relation = isSameStructure
-                ? existingRelation
-                : await CreateComponentRelationAsync(component, slotName, parentId, order, cancellationToken);
-
-            // 子要素の処理
-            var existingChildrenMap = BuildExistingChildrenMap(existingRelation);
-
-            if (newNode.Children != null && newNode.Children.Count > 0)
-            {
-                await MergeChildrenAsync(newNode, relation, existingRelation, existingChildrenMap, cancellationToken);
-            }
-            else if (existingRelation != null)
-            {
-                // 子要素がない場合で、既存の子がある場合のみ削除
-                await DeleteAllChildrenAsync(existingRelation, cancellationToken);
-            }
-
-            return relation;
-        }
-
-        /// <summary>
-        /// 新しいComponentRelationを作成
-        /// </summary>
-        private async Task<ComponentRelation> CreateComponentRelationAsync(
-            Component component,
-            string slotName,
-            Guid? parentId,
-            int order,
-            CancellationToken cancellationToken)
-        {
-            var relation = new ComponentRelation
-            {
-                Id = Guid.NewGuid(),
-                ParentId = parentId,
-                ChildId = component.Id,
-                SlotName = slotName,
-                Order = order,
-                IsPriority = false
-            };
-
-            await _componentRelationRepository.CreateAsync(relation, cancellationToken);
-            return relation;
-        }
-
-        /// <summary>
-        /// ComponentRelation と その子孫をすべて削除
-        /// </summary>
-        private async Task DeleteComponentRelationTree(ComponentRelation relation, CancellationToken cancellationToken)
-        {
-            await DeleteComponentRelationTreeWithTracking(relation, new HashSet<Guid>(), cancellationToken);
-        }
-
-        /// <summary>
-        /// ComponentRelationツリーの削除（削除済みIDを追跡）
-        /// </summary>
-        private async Task DeleteComponentRelationTreeWithTracking(
-            ComponentRelation relation,
-            HashSet<Guid> deletedIds,
-            CancellationToken cancellationToken)
-        {
-            if (relation == null || deletedIds.Contains(relation.Id))
-            {
-                return;
-            }
-
-            deletedIds.Add(relation.Id);
-
-            // このComponentRelationの子要素（relation.Child の下の要素）を再帰的に削除
-            if (relation.Child?.Children != null && relation.Child.Children.Count > 0)
-            {
-                var childrenToDelete = relation.Child.Children.ToList();
-                foreach (var child in childrenToDelete)
-                {
-                    await DeleteComponentRelationTreeWithTracking(child, deletedIds, cancellationToken);
-                }
-            }
-
-            // このComponentRelationを削除
-            await _componentRelationRepository.DeleteAsync(relation, cancellationToken);
-        }
-
-        /// <summary>
-        /// ComponentIdを検証してComponentを取得
-        /// </summary>
-        private async Task<Component> ExtractComponentAsync(string componentId, CancellationToken cancellationToken)
-        {
-            if (!Guid.TryParse(componentId, out var parsedId))
-            {
-                return null;
-            }
-
-            return await _componentRepository.GetByIdAsync(parsedId, cancellationToken);
-        }
-
-        /// <summary>
-        /// 子要素のキーを生成（PackageId + Name + SlotName）
-        /// </summary>
-        private static string CreateChildRelationKey(string packageId, string name, string slotName)
-        {
-            return $"{packageId}_{name}_{slotName}";
-        }
-
-        /// <summary>
-        /// 既存のComponentRelationが新しいツリーと同じ構造か判定
-        /// </summary>
-        private static bool IsRelationUnchanged(
-            ComponentRelation existingRelation,
-            Component newComponent,
-            string newSlotName,
-            Guid? parentId,
-            int order)
-        {
-            if (existingRelation?.Child == null)
-            {
-                return false;
-            }
-
-            return existingRelation.Child.PackageId == newComponent.PackageId &&
-                   existingRelation.Child.Name == newComponent.Name &&
-                   existingRelation.SlotName == newSlotName &&
-                   existingRelation.ParentId == parentId &&
-                   existingRelation.Order == order;
-        }
-
-        /// <summary>
-        /// 既存の子要素をキー別にマッピング
-        /// </summary>
-        private static Dictionary<string, ComponentRelation> BuildExistingChildrenMap(ComponentRelation existingRelation)
-        {
-            var map = new Dictionary<string, ComponentRelation>();
-
-            if (existingRelation?.Child?.Children == null)
-            {
-                return map;
-            }
-
-            foreach (var child in existingRelation.Child.Children)
-            {
-                var key = CreateChildRelationKey(
-                    child.Child?.PackageId.ToString() ?? "",
-                    child.Child?.Name ?? "",
-                    child.SlotName);
-
-                map[key] = child;
-            }
-
-            return map;
-        }
-
-        /// <summary>
-        /// 新しい子要素を処理・マージ
-        /// </summary>
-        private async Task MergeChildrenAsync(
-            ComponentTreeNode newNode,
-            ComponentRelation relation,
-            ComponentRelation existingRelation,
-            Dictionary<string, ComponentRelation> existingChildrenMap,
-            CancellationToken cancellationToken)
-        {
-            if (newNode?.Children == null || newNode.Children.Count == 0)
-            {
-                return;
-            }
-
-            var processedChildKeys = new HashSet<string>();
-            int childOrder = 0;
-
-            foreach (var newChild in newNode.Children)
-            {
-                var childComponent = await ExtractComponentAsync(newChild.ComponentId, cancellationToken);
-                if (childComponent == null)
-                {
-                    continue;
-                }
-
-                var childSlotName = newChild.SlotName ?? "Main";
-                var childKey = CreateChildRelationKey(childComponent.PackageId.ToString(), childComponent.Name, childSlotName);
-                processedChildKeys.Add(childKey);
-
-                var existingChild = existingChildrenMap.TryGetValue(childKey, out var found) ? found : null;
-
-                // 再帰的にマージ
-                await MergeComponentRelationTree(newChild, existingChild, relation.ChildId, childOrder++, cancellationToken);
-            }
-
-            // 処理されなかった子要素は削除対象
-            await DeleteObsoleteChildrenAsync(existingChildrenMap, processedChildKeys, cancellationToken);
-        }
-
-        /// <summary>
-        /// 処理されなかった子要素を削除
-        /// </summary>
-        private async Task DeleteObsoleteChildrenAsync(
-            Dictionary<string, ComponentRelation> existingChildrenMap,
-            HashSet<string> processedChildKeys,
-            CancellationToken cancellationToken)
-        {
-            var obsoleteKeys = existingChildrenMap.Keys.Where(k => !processedChildKeys.Contains(k));
-
-            foreach (var key in obsoleteKeys)
-            {
-                var relationToDelete = existingChildrenMap[key];
-                await DeleteComponentRelationTree(relationToDelete, cancellationToken);
-            }
-        }
-
-        /// <summary>
-        /// 既存の子要素をすべて削除
-        /// </summary>
-        private async Task DeleteAllChildrenAsync(ComponentRelation existingRelation, CancellationToken cancellationToken)
-        {
-            if (existingRelation == null)
-            {
-                return;
-            }
-
-            // ParentIdがexistingRelation.ChildIdである全てのRelationを削除
-            var parentId = existingRelation.ChildId;
-
-            // データベースから削除対象の子要素を全て取得
-            var allRelations = await _componentRelationRepository.GetAllAsync(cancellationToken);
-            var childrenToDelete = allRelations
-                .Where(r => r.ParentId == parentId)
-                .ToList();
-
-            // 各子要素に対してツリー削除を実行
-            foreach (var childRelation in childrenToDelete)
-            {
-                await DeleteComponentRelationTree(childRelation, cancellationToken);
-            }
-        }
 
         /// <summary>
         /// コンポーネント情報とそのスロット情報を取得
@@ -994,7 +460,6 @@ namespace Micon.CMS.Controllers
                         var componentType = AssemblyService.GetType(component.PackageId, component.Name);
                         if (componentType != null)
                         {
-                            Console.WriteLine($"[DEBUG] Analyzing slots for component: {component.Name}");
                             // 特定のコンポーネント型のスロット情報を取得
                             // assemblyBasePath を指定しない（DLL の埋め込みリソースから読む）
                             var slots = _slotAnalyzerService.AnalyzeComponentSlots(componentType, null);
@@ -1003,26 +468,21 @@ namespace Micon.CMS.Controllers
                                 .Distinct()
                                 .ToList();
 
-                            Console.WriteLine($"[DEBUG] Found {slotNames.Count} slots: {string.Join(", ", slotNames)}");
                             dto.Slots = slotNames;
                         }
                         else
                         {
-                            Console.WriteLine($"[DEBUG] Component type not found for: {component.Name}");
                             dto.Slots = new List<string>();
                         }
                     }
                     else
                     {
-                        Console.WriteLine($"[DEBUG] Assembly not found for PackageId: {component.PackageId}");
                         dto.Slots = new List<string>();
                     }
                 }
                 catch (Exception ex)
                 {
                     // スロット取得に失敗した場合は空リスト
-                    Console.WriteLine($"[ERROR] Error analyzing slots for component {component.Name}: {ex.Message}");
-                    Console.WriteLine($"[ERROR] Stack trace: {ex.StackTrace}");
                     dto.Slots = new List<string>();
                 }
 
@@ -1057,14 +517,5 @@ namespace Micon.CMS.Controllers
     {
         public required string PageTemplateId { get; set; }
         public List<ComponentTreeNode>? Tree { get; set; }
-    }
-
-    public class ComponentTreeNode
-    {
-        public required string ComponentId { get; set; }
-        public required string ComponentName { get; set; }
-        public required string PackageId { get; set; }
-        public required string SlotName { get; set; }
-        public List<ComponentTreeNode> Children { get; set; } = new List<ComponentTreeNode>();
     }
 }
